Moves base plugin into prelude
