+++ lib/staticky/resource.rb
@@ -2,69 +2,6 @@
 
 module Staticky
   class Resource
-    module ClassMethods
-      def new(**env)
-        super().tap do |resource|
-          env.each do |key, value|
-            resource.send(:"#{key}=", value)
-          end
-        end
-      end
-    end
-
-    module InstanceMethods
-      def filepath
-        destination.join(basename)
-      end
-
-      def read
-        filepath.read
-      end
-
-      def basename
-        root? ? "index.html" : "#{url}.html"
-      end
-
-      def root?
-        url == "/"
-      end
-
-      def uri
-        raise ArgumentError, "url is required" unless defined?(@uri)
-
-        @uri
-      end
-
-      def destination
-        @destination ||= Staticky.build_path
-      end
-
-      def destination=(destination)
-        @destination = Pathname(destination)
-      end
-
-      def url
-        raise ArgumentError, "url is required" unless defined?(@url)
-
-        @url
-      end
-
-      def url=(url)
-        @url = url
-        @uri = parse_url(url)
-      end
-
-      private
-
-      def parse_url(url)
-        URI(url).tap do |uri|
-          uri.path = "/#{uri.path}" unless uri.path.start_with?("/")
-        end
-      rescue URI::InvalidURIError => e
-        raise ArgumentError, e.message
-      end
-    end
-
     def self.plugin(plugin, *, &block)
       plugin = Resources::Plugins.load_plugin(plugin) if plugin.is_a?(Symbol)
 
@@ -78,7 +15,7 @@ module Staticky
       plugin.configure(self, *, &block) if plugin.respond_to?(:configure)
     end
 
-    plugin self
+    plugin :prelude
     plugin :phlex
   end
 end
