+++ lib/staticky/resources/plugins.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Staticky
   module Resources
     module Plugins
